Fix Pinboard JSON duplicate timestamps error
If the JSON exported by Pinboard contains duplicate timestamps, Python
returns a TypeError exception:

`TypeError: argument of type 'float' is not iterable`

This is because `time.mktime()` returns a floating point number.
Encasing `time.mktime()` in `str()` fixes the data type not being
iterable.

`time.mktime()` has also been encased in `int()` to remove the
unnecessary decimal value (`.0`) that gets returned for each time value,
and to keep the script consistent with the other export functions.