Update timestamp in index after a new snapshot is made.
After a recent archive run that was somehow borked, I found archive in a
state where index was linking to a directory which didn't exist in archive.
When I tried to archive the link again, it was successful but the index
entry continued to point to the old timestamp, therefore the old
(non-existent) directory.

I think it makes sense to update the timestamp after making a new snapshot,
so that index points to the latest files.